silenciar funciona

+++ adminUsuarios.php
@@ -37,8 +37,13 @@ include_once("usuarios.php");
 							<ul class="dropdown-menu" aria-labelledby="dropdownMenuButton1">';
 							echo '<li><a class="dropdown-item" href="modificarUsuario.php?formModificar&id='.$usuarios[$i]['id'].'"><i class="fas fa-edit me-1"></i> Editar</a></li>';
 							echo '<li><a class="dropdown-item" href="usuarios.php?aksi=delete&id='.$usuarios[$i]['id'].'" onclick="return confirm(\'Esta seguro de borrar los datos de '.$usuarios[$i]['nombre'].'?\')"><i class="fas fa-trash-alt me-1"></i> Borrar</a></li>';
-							echo '<li><a class="dropdown-item" href="usuarios.php?aksi=silence&id='.$usuarios[$i]['id'].'" onclick="return confirm(\'Esta seguro de silenciar a '.$usuarios[$i]['nombre'].'?\')"><i class="fas fa-comment-slash me-1"></i>Silenciar</a></li>';
-							echo '</ul>';
+							if($usuarios[$i]['inactivar_coment']==1){
+                                echo '<li><a class="dropdown-item" href="usuarios.php?aksi=disilence&id='.$usuarios[$i]['id'].'" onclick="return confirm(\'Esta seguro de desilenciar a '.$usuarios[$i]['nombre'].'?\')"><i class="fas fa-comment-dots me-1"></i>Desilenciar</a></li>';
+                            }
+                            else{
+                                echo '<li><a class="dropdown-item" href="usuarios.php?aksi=silence&id='.$usuarios[$i]['id'].'" onclick="return confirm(\'Esta seguro de silenciar a '.$usuarios[$i]['nombre'].'?\')"><i class="fas fa-comment-slash me-1"></i>Silenciar</a></li>';
+                            }
+                            echo '</ul>';
 							echo '</div>';
 							echo '</td>';
                             echo '</tr>';
